requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.107.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
Flask==3.0.0
Flask-CORS==4.0.0
cryptography==41.0.7
reportlab==4.0.9
pypandoc==1.12
python-docx==1.1.0
icalendar==5.0.11
Werkzeug==3.0.1
Pillow==10.2.0
lxml==5.1.0
markdown==3.5.2